# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vehicle Transport Automation",
    "description": "Control Panel for Email-to-ClickUp Pipeline with ML Training Support",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/review/{run_id}": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Get Review For Run",
        "description": "Get all review items for an extraction run.\n\nReturns the extraction run info and all review items to be reviewed.",
        "operationId": "get_review_for_run_api_review__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/{run_id}/item/{item_id}": {
      "put": {
        "tags": [
          "Review"
        ],
        "summary": "Update Review Item",
        "description": "Update a single review item.\n\nMark as correct (is_match_ok=true) or provide a corrected value.",
        "operationId": "update_review_item_api_review__run_id__item__item_id__put",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/submit": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Submit Review",
        "description": "Submit review corrections for an extraction run.\n\nThis endpoint:\n1. Updates all review items with corrections\n2. Creates TrainingExamples from the corrected data\n3. Optionally marks the run as reviewed\n\nTrainingExamples are used for future ML model training.",
        "operationId": "submit_review_api_review_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/{run_id}/approve": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Approve Run",
        "description": "Mark an extraction run as reviewed/approved.\n\nUse this after all items have been reviewed.",
        "operationId": "approve_run_api_review__run_id__approve_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/training-examples/": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "List Training Examples",
        "description": "List training examples for ML training.\n\nFilter by auction type, field, or correctness.",
        "operationId": "list_training_examples_api_review_training_examples__get",
        "parameters": [
          {
            "name": "auction_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auction Type Id"
            }
          },
          {
            "name": "field_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Field Key"
            }
          },
          {
            "name": "is_correct",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Correct"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingExamplesListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/training-examples/export": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Export Training Data",
        "description": "Export training examples for ML training.\n\nReturns JSONL or CSV format suitable for fine-tuning.",
        "operationId": "export_training_data_api_review_training_examples_export_get",
        "parameters": [
          {
            "name": "auction_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by auction type",
              "title": "Auction Type Id"
            },
            "description": "Filter by auction type"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format: jsonl or csv",
              "default": "jsonl",
              "title": "Format"
            },
            "description": "Export format: jsonl or csv"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewItemUpdate": {
        "properties": {
          "corrected_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Value",
            "description": "Corrected value (if different from predicted)"
          },
          "is_match_ok": {
            "type": "boolean",
            "title": "Is Match Ok",
            "description": "True if predicted value is correct"
          },
          "export_field": {
            "type": "boolean",
            "title": "Export Field",
            "description": "Include this field in export",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "is_match_ok"
        ],
        "title": "ReviewItemUpdate",
        "description": "Update a single review item."
      },
      "ReviewRunResponse": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "document_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Filename"
          },
          "auction_type_id": {
            "type": "integer",
            "title": "Auction Type Id"
          },
          "auction_type_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auction Type Code"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "extraction_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Score"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "reviewed_count": {
            "type": "integer",
            "title": "Reviewed Count"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "document_id",
          "auction_type_id",
          "status",
          "items",
          "reviewed_count",
          "total_count"
        ],
        "title": "ReviewRunResponse",
        "description": "Response for a review run with all items."
      },
      "ReviewSubmitRequest": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id",
            "description": "Extraction run ID"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemUpdate"
            },
            "type": "array",
            "title": "Items",
            "description": "Updated review items"
          },
          "mark_as_reviewed": {
            "type": "boolean",
            "title": "Mark As Reviewed",
            "description": "Mark run as reviewed after submit",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "items"
        ],
        "title": "ReviewSubmitRequest",
        "description": "Submit review corrections for a run."
      },
...
      "TrainingExamplesListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrainingExampleResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TrainingExamplesListResponse",
        "description": "Response for training examples list."
      },
...
    }
  }
}